## Overview

Noticed that the eval of very simple watch expressions was a bit slower than expected and dug into it, found a quick fix and reduced the watch expression time for this particular case by 90%; should have similar speedups for any expressions with deep transitive dependencies.

I ran a test locally with a single top-level example expression that had a deep dependency tree, and a single watch expression.

Here's how the timings compare on a couple runs (after runtime warmup)

```
trunk:
Finished Collect deps in 0.951952s (cpu), 0.951724s (system)
Finished Collect deps in 0.956545s (cpu), 0.958644s (system)

this branch:
Finished Collect deps in 0.122703s (cpu), 0.122523s (system)
Finished Collect deps in 0.14749s (cpu), 0.163158s (system)
```

## Implementation notes

There was a subtlety in the optimization of recursive dependency search in the runtime which meant that tree pruning wasn't shared amongst multiple branches. I swapped it to a state monad rather than explicit binding passing to ensure the 'seen' Set is propagated everywhere it needs to go.

## Test coverage

Local testing using `UNISON_DEBUG=TIMING`

## Loose ends

Could probably do even better by only collecting dependencies of the watch expressions themselves, I don't think we need the dependencies of parts of the file that we don't actually touch. Would that work @dolio ?

I'd also be curious if we could delay the code lookup until execution time so we only load terms that are literally required by the exact evaluation of the watch expression, but I'm guessing this would be a pain to write because we'd need separate codepaths for the watch expr eval vs the version that runs from a file or .uc or w/e